Split JS to files
